% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_power_sensitivity.R
\name{analyze_power_sensitivity}
\alias{analyze_power_sensitivity}
\title{Analyze Power Sensitivity to VDD and Frequency}
\usage{
analyze_power_sensitivity(alpha_fixed, C_fixed, V_DD_range, f_range)
}
\arguments{
\item{alpha_fixed}{The fixed average switching activity factor (e.g., average alpha).}

\item{C_fixed}{The fixed total capacitance (in Farads, F).}

\item{V_DD_range}{A numeric vector of supply voltage values (in Volts, V) to test.}

\item{f_range}{A numeric vector of clock frequency values (in Hertz, Hz) to test.}
}
\value{
A matrix where rows correspond to VDD values and columns to frequency values.
Each element contains the estimated dynamic power (in Watts, W).
}
\description{
Performs a sensitivity analysis by calculating dynamic power over a range of
VDD and frequency values, by fixing switching activity and capacitance.
}
\examples{
VDDs <- seq(0.8, 1.2, by = 0.1) # 0.8V to 1.2V
Freqs <- seq(0.5e9, 2e9, length.out = 4) # 0.5 GHz to 2 GHz
power_matrix <- analyze_power_sensitivity(0.15, 60e-12, VDDs, Freqs)
print(power_matrix)

}
