% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dynamic_function.R
\name{estimate_dynamic_power}
\alias{estimate_dynamic_power}
\title{Estimate CMOS Dynamic Power Dissipation}
\usage{
estimate_dynamic_power(alpha, C, V_DD, f)
}
\arguments{
\item{alpha}{The average switching activity factor (0 to 1). If calculating for a word,
it is the total number of transitions divided by the total number of bits.}

\item{C}{The total capacitance being switched (in Farads, F).}

\item{V_DD}{The supply voltage (in Volts, V).}

\item{f}{The operating clock frequency (in Hertz, Hz).}
}
\value{
A numeric value representing the estimated dynamic power (in Watts, W).
}
\description{
Calculates the estimated dynamic power consumption (\eqn{P_{dyn}})
based on the standard CMOS formula: \eqn{P_{dyn} = 0.5 \cdot \alpha \cdot C_L \cdot V_{DD}^2 \cdot f}.
}
\examples{
# Assuming low activity (alpha=0.1), typical capacitance, voltage, and frequency
alpha_val <- 0.1
C_L_val <- 50e-12    # 50 pF
V_DD_val <- 1.0     # 1.0 V
f_val <- 1e9        # 1 GHz
P_dyn <- estimate_dynamic_power(alpha_val, C_L_val, V_DD_val, f_val)
print(P_dyn)

}
