% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_benchmark_power.R
\name{plot_benchmark_power}
\alias{plot_benchmark_power}
\title{Plot Power Benchmark Comparison and Save as PDF}
\usage{
plot_benchmark_power(
  data_stream_original,
  data_stream_optimized,
  C_L_fixed,
  V_DD_fixed,
  f_fixed,
  output_filename_pdf = "benchmark_comparison"
)
}
\arguments{
\item{data_stream_original}{A list of binary vectors (0s and 1s) representing the original sequence of words.}

\item{data_stream_optimized}{A list of binary vectors (0s and 1s) representing the optimized sequence of words.}

\item{C_L_fixed}{The fixed total load capacitance (in Farads, F).}

\item{V_DD_fixed}{The fixed supply voltage (in Volts, V).}

\item{f_fixed}{The fixed clock frequency (in Hertz, Hz).}

\item{output_filename_pdf}{The path and filename (ending in .pdf) to save the plot.
Standardly, it is "benchmark_comparison.pdf"}
}
\value{
The ggplot object (also saves a PDF file and an MP4 file).
}
\description{
Calls benchmark_power to calculate the power reduction, generates
a static ggplot bar chart, and saves it to a PDF file.
}
\examples{
\dontrun{
D_orig <- list(c(0,1,0,0), c(1,1,0,1), c(1,1,1,0))
D_opt  <- list(c(1,0,0,1), c(0,0,1,1), c(1,0,0,0))
C_L <- 50e-12; V_DD <- 1.0; f <- 1e9
power_plot <- plot_benchmark_power(D_orig, D_opt, C_L, V_DD, f)
}

}
