% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_power_sensitivity.R
\name{plot_power_sensitivity}
\alias{plot_power_sensitivity}
\title{Plot and Animate Power Sensitivity Analysis}
\usage{
plot_power_sensitivity(
  alpha_fixed,
  C_L_fixed,
  V_DD_range,
  f_range,
  base_output_filename_base = "power_sensitivity"
)
}
\arguments{
\item{alpha_fixed}{The fixed average switching activity factor.}

\item{C_L_fixed}{The fixed total load capacitance (in Farads, F).}

\item{V_DD_range}{A numeric vector of supply voltage values (in Volts, V) to test.}

\item{f_range}{A numeric vector of clock frequency values (in Hertz, Hz) to test.}

\item{base_output_filename_base}{The base path and filename (e.g., "sensitivity_")
The function will append "_VDD.pdf", "_Freq.pdf",
"_VDD.mp4", and "_Freq.mp4". Standardly, it is "power_sensitivity"}
}
\value{
A list containing the two ggplot objects and the two gganimate objects (also saves PDF and MP4 files).
}
\description{
Calls analyze_power_sensitivity, transforms the data, and generates two
static ggplot objects (P_dyn vs VDD, P_dyn vs Freq), saving each one to a separate PDF file.
Also generates two animated MP4 videos showing the sensitivity.
}
\examples{
\dontrun{
VDDs <- seq(0.8, 1.2, length.out = 10)
Freqs <- seq(0.5e9, 2.5e9, length.out = 10) # Augmenté le nombre de points pour meilleure animation
sensitivity_outputs <- plot_power_sensitivity(0.15, 60e-12, VDDs, Freqs)
}
}
