% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vlsi_data_stream}
\alias{vlsi_data_stream}
\title{Flux de données binaires VLSI pour tests de puissance}
\format{
Une liste contenant 2000 vecteurs numériques de longueur 32.
}
\source{
Simulation aléatoire via \code{data-raw/vlsi_data_stream.R}
}
\usage{
vlsi_data_stream
}
\description{
Un dataset simulé représentant 2000 cycles d'un bus de données 32 bits.
Chaque élément de la liste est un vecteur numérique de 0 et 1.
}
\keyword{datasets}
